moved Poll.vue and QuizEdit.vue to admin directory and modified router.js to reflect the change
